item 디스플레이 및 북마크 관련 컴포넌트 세분화
- ItemViewerBox 컴포넌트 생성 : item 관련 정보 받아서 렌더링 전담하는 컴포넌트
- BookmarkButton 컴포넌트 생성 : Bookmark 상태 전담하는 컴포넌트
